headers: use u64 (not usize) in Payload structs for wasm32 compatibility
`Header<T>: Serializable` serializes the `#[repr(C)]` payload as a raw
`mem::size_of` byte copy. On 32-bit targets (e.g. wasm32) `usize` is 4
bytes, but the on-disk format was written on a 64-bit host with 8-byte
usize, so every field after the first one is read from the wrong bytes
— including the `flags` field that gates the SDSL check.

Symptom for downstream wasm consumers (e.g. building gbz-base for
wasm32-wasip1 to load a .gbz in a browser): load fails with the
spurious error "GBWTHeader: SDSL format is not supported" against any
valid modern .gbz file. The bytes that should be `flags` happen to
land on the high bits of `offset`, and the SDSL bit looks set.

Changing the affected fields to explicit `u64` keeps the on-disk
layout identical on every target. Accessors that already expose
these values as `usize` (Gbwt::sequences/len/alphabet_size/...,
Metadata::samples/haplotypes/contigs, Sequences::nodes) cast at the
boundary; the builder side casts the other direction. No file format
change; the on-disk bytes are exactly what 64-bit builds already
write.

+++ src/gbwt.rs
@@ -108,7 +108,7 @@ impl GBWT {
     /// Returns the total length of the sequences in the index.
     #[inline]
     pub fn len(&self) -> usize {
-        self.header.payload().size
+        self.header.payload().size as usize
     }
 
     /// Returns `true` if the index is empty.
@@ -120,19 +120,19 @@ impl GBWT {
     /// Returns the number of sequences in the index.
     #[inline]
     pub fn sequences(&self) -> usize {
-        self.header.payload().sequences
+        self.header.payload().sequences as usize
     }
 
     /// Returns the size of the alphabet.
     #[inline]
     pub fn alphabet_size(&self) -> usize {
-        self.header.payload().alphabet_size
+        self.header.payload().alphabet_size as usize
     }
 
     /// Returns the alphabet offset for the effective alphabet.
     #[inline]
     pub fn alphabet_offset(&self) -> usize {
-        self.header.payload().offset
+        self.header.payload().offset as usize
     }
 
     /// Returns the size of the effective alphabet.
@@ -424,7 +424,7 @@ impl Serialize for GBWT {
             return Err(Error::new(ErrorKind::InvalidData, "GBWT: Invalid metadata flag in the header"));
         }
         if let Some(meta) = metadata.as_ref() && meta.has_path_names() {
-            let expected = if header.is_set(GBWTPayload::FLAG_BIDIRECTIONAL) { header.payload().sequences / 2 } else { header.payload().sequences };
+            let expected = if header.is_set(GBWTPayload::FLAG_BIDIRECTIONAL) { (header.payload().sequences / 2) as usize } else { header.payload().sequences as usize };
             if meta.paths() > 0 && meta.paths() != expected {
                 return Err(Error::new(ErrorKind::InvalidData, "GBWT: Invalid path count in the metadata"));
             }
